Start refactoring how value blocks are used
The current implementation for value blocks does a poor job of handling
fields that aren't 4 bytes. It also isn't really meant to be used as a
read/write store, so you have to store the values separately, then write
them all once you are ready to send it. Managing the field masks is also
a headache.

A new approach takes some inspiration from Gophercraft by just using a
struct to represent the value block. Setters mark the fields as dirty to
make partial updates easier later on. Unlike Gophercraft which uses
reflection heavily at runtime, I want to instead inspect the structs
once on startup to get the field offsets and sizes. The dirty flags map
to the field index in the struct, which means the flag can be used to
lookup the offset/size. This removes the need for the FieldMask list and
allows offsets to be inferred based on the field layout. The obvious
consequence of this is that the data structure is fixed, however it's
possible to add additional data at the end of the struct that won't
impact the offsets.